use int as default type

+++ src/sdpi/lapack_interface.c
@@ -62,8 +62,8 @@
 #include <cblas.h>
 typedef blasint LAPACKINTTYPE;
 #else
-/* otherwise we use 64 bit integers as the base type because SDPA 7.4.0 uses this by default */
-typedef long long int LAPACKINTTYPE;
+/* otherwise we use int as a default */
+typedef int LAPACKINTTYPE;
 #endif
 
 /** Checks if a BMSallocMemory-call was successfull, otherwise returns SCIP_NOMEMORY */
